requirements.txt
streamlit
google-generativeai
serpapi
folium
streamlit-folium
pandas
requests
pillow